requirements.txt
python-dotenv

# --- Web UI ---
#streamlit>=1.33.0
streamlit

# --- LLM & LangChain ---
#langchain>=0.2.0
#langchain-openai>=0.1.6
#langchain-core>=0.2.0
#pydantic>=2.6.0
langchain
langchain-openai
langchain-core
pydantic

# --- 日付処理 ---
#python-dateutil>=2.9.0
#dateparser>=1.2.0
#zoneinfo; python_version < "3.9"
python-dateutil
dateparser

# --- 補助 ---
#tqdm>=4.66.0
tqdm

# --- テスト ---
#pytest>=8.0.0
pytest

# --- 型サポート（任意だが推奨）---
#typing-extensions>=4.9.0
typing-extensions